// ts/react/common/types/bedev2Types.ts
export enum TContentType {
	Game = "Game",
	CatalogAsset = "CatalogAsset",
	CatalogBundle = "CatalogBundle",
}
// ...
export type TOmniRecommendationGame = {
	contentType: TContentType.Game;
	contentId: number;
	contentMetadata: Record<string, string>;
};
// ...
type TSharedGameSort = {
	topic: string;
	topicId: number;
	treatmentType:
		| TTreatmentType.Carousel
		| TTreatmentType.SortlessGrid
		| TTreatmentType.InterestGrid;
	subtitle?: string;
	topicLayoutData?: TTopicLayoutData;
};

export type TOmniRecommendationSduiSort = {
	topicId: number;
	treatmentType: TTreatmentType.Sdui;
	numberOfRows?: number;

	feedItemKey: string;
};

export type TOmniRecommendationGameSort = TSharedGameSort & {
	recommendationList: TOmniRecommendationGame[];
	numberOfRows?: number;
};
// ...
export type TExploreApiGameSort = TSharedGameSort & {
	games: TGameData[];
	sortId: string;
	contentType: string;
	nextPageToken: string;
	gameSetTargetId?: number;
	appliedFilters?: string;
};
// ...
export type TGameSort = TExploreApiGameSort | TOmniRecommendationGameSort;