// ts/react/omniFeed/GameGridFeedItem.4d1f140ec42445865b30144d9a74e13eac40b98d8986a8c488a8ccc152b82606-GameCarousel.js.tsx
import React, { useEffect, useCallback, useMemo } from "react";
import { WithTranslationsProps } from "react-utilities";
import { TGameData, TGetFriendsResponse } from "../common/types/bedev1Types";
import {
	TContentType,
	TGameSort,
	TOmniRecommendationGame,
} from "../common/types/bedev2Types";
import HomePageGridDiscoveryApi from "../homePage/discoveryApi/HomePageGridDiscoveryApi";
import bedev2Services from "../common/services/bedev2Services";
import { homePage } from "../common/constants/configConstants";
import { useContentMetadata } from "./utils/contentMetadataContextProvider";
import { hydrateOmniRecommendationGames } from "./utils/gameSortUtils";
import { usePageSession } from "../common/utils/PageSessionContext";

type THomePageDiscoveryApiProps = {
	translate: WithTranslationsProps["translate"];
	sort: TGameSort;
	positionId: number;
	itemsPerRow: number | undefined;
	startingRow: number | undefined;
	recommendations: TOmniRecommendationGame[];
	friendsPresenceData: TGetFriendsResponse[];
	isExpandHomeContentEnabled?: boolean;
	isNewSortHeaderEnabled?: boolean;
	hiddenUniverses?: Set<number>;
	setHiddenUniverses?: React.Dispatch<React.SetStateAction<Set<number>>>;
};

const { sortlessGridMaxTilesMetadataToFetch } = homePage;

export const GameGridFeedItem = ({
	translate,
	sort,
	positionId,
	itemsPerRow,
	startingRow,
	recommendations,
	friendsPresenceData,
	isExpandHomeContentEnabled,
	isNewSortHeaderEnabled,
	hiddenUniverses,
	setHiddenUniverses,
}: THomePageDiscoveryApiProps): JSX.Element | null => {
	const homePageSessionInfo = usePageSession();

	const { contentMetadata, appendContentMetadata } = useContentMetadata();

	const fetchAdditionalMetadata = useCallback(() => {
		const gamesWithoutMetadata = recommendations.filter(
			({
				contentType,
				contentId,
			}: {
				contentType: TContentType.Game;
				contentId: number;
			}) => !contentMetadata?.[contentType]?.[contentId],
		);

		if (gamesWithoutMetadata.length > 0) {
			bedev2Services
				.getOmniRecommendationsMetadata(
					gamesWithoutMetadata.slice(0, sortlessGridMaxTilesMetadataToFetch),
					homePageSessionInfo,
				)
				.then((additionalMetadata) =>
					appendContentMetadata(additionalMetadata.contentMetadata),
				)
				.catch(() => {
					// empty catch block, as this is not blocking. we will show the games we already have metadata for
				});
		}
	}, [
		recommendations,
		homePageSessionInfo,
		contentMetadata,
		appendContentMetadata,
	]);

	useEffect(() => {
		fetchAdditionalMetadata();
	}, [fetchAdditionalMetadata]);

	const gridData: TGameData[] = useMemo(() => {
		return hydrateOmniRecommendationGames(recommendations, contentMetadata);
	}, [recommendations, contentMetadata]);

	if (gridData?.length === 0) {
		return null;
	}

	return (
		<HomePageGridDiscoveryApi
			key={sort.topic}
			sort={sort}
			gameData={gridData}
			translate={translate}
			positionId={positionId}
			itemsPerRow={itemsPerRow}
			startingRow={startingRow}
			friendsPresence={friendsPresenceData}
			componentType={sort.topicLayoutData?.componentType}
			playerCountStyle={sort.topicLayoutData?.playerCountStyle}
			playButtonStyle={sort.topicLayoutData?.playButtonStyle}
			hoverStyle={sort.topicLayoutData?.hoverStyle}
			isSponsoredFooterAllowed={
				sort.topicLayoutData?.isSponsoredFooterAllowed === "true"
			}
			hideTileMetadata={sort.topicLayoutData?.hideTileMetadata === "true"}
			isExpandHomeContentEnabled={isExpandHomeContentEnabled}
			isNewSortHeaderEnabled={isNewSortHeaderEnabled}
			enableExplicitFeedback={
				sort.topicLayoutData?.enableExplicitFeedback === "true"
			}
			hiddenUniverses={hiddenUniverses}
			setHiddenUniverses={setHiddenUniverses}
		/>
	);
};

GameGridFeedItem.defaultProps = {
	isExpandHomeContentEnabled: undefined,
	isNewSortHeaderEnabled: undefined,
};

export default GameGridFeedItem;
